forgot to add a use example in update
earlier i presented some example methods that changes the activity status variables, but they weren't in update. so here's me adding it to it.

it's a little bit of an expensive example because i'm checking strings. but you get the idea. i reccomend making enums to check what kind of scene you are in.